+++ okdata/cli/commands/teams/questions.py
@@ -9,6 +9,12 @@ class NoTeamError(Exception):
     pass
 
 
+class NoTeamMembersError(Exception):
+    """Raised when a team does not have any members."""
+
+    pass
+
+
 def q_team(team_client, my=True):
     def _team_choices():
         teams = team_client.get_teams(
@@ -74,4 +80,42 @@ def q_attribute_value():
         "name": "attribute_value",
         "message": "New value",
         "when": lambda x: x["attribute"][0] != "name",
+    }
+
+
+def q_username():
+    return {
+        **common_style,
+        "type": "text",
+        "name": "username",
+        "message": "Username (ident)",
+        "validate": lambda t: bool(t) or "Can't be empty",
+    }
+
+
+def q_members(team_client):
+    def _team_member_choices(team_id):
+        members = team_client.get_team_members(team_id)
+
+        if not members:
+            raise NoTeamMembersError
+
+        return [
+            Choice(
+                f"{m['name']} ({m['username']})" if m["name"] else m["username"],
+                m["username"],
+            )
+            for m in sorted(
+                members,
+                key=lambda u: (not u["name"], u["name"] or "", u["username"]),
+            )
+        ]
+
+    return {
+        **common_style,
+        "type": "checkbox",
+        "name": "usernames",
+        "message": "Member(s)",
+        "choices": lambda x: _team_member_choices(x["team_id"]),
+        "validate": lambda x: bool(x) or "Select at least one team member",
     }