+++ okdata/cli/commands/teams/teams.py
@@ -5,7 +5,12 @@ from requests.exceptions import HTTPError
 
 from okdata.cli.command import BASE_COMMAND_OPTIONS, BaseCommand
 from okdata.cli.commands.teams.questions import NoTeamError
-from okdata.cli.commands.teams.wizards import edit_team_wizard, list_members_wizard
+from okdata.cli.commands.teams.wizards import (
+    add_member_wizard,
+    edit_team_wizard,
+    list_members_wizard,
+    remove_member_wizard,
+)
 from okdata.cli.output import create_output
 
 
@@ -16,12 +21,16 @@ Usage:
   okdata teams ls [--my] [options]
   okdata teams edit [options]
   okdata teams list-members [--my] [options]
+  okdata teams add-member [options]
+  okdata teams remove-member [options]
 
 Examples:
   okdata teams ls
   okdata teams ls --my
   okdata teams edit
   okdata teams list-members --my
+  okdata teams add-member
+  okdata teams remove-member
 
 Options:{BASE_COMMAND_OPTIONS}
     """
@@ -37,6 +46,10 @@ Options:{BASE_COMMAND_OPTIONS}
             self.edit()
         if self.cmd("list-members"):
             self.list_members(self.opt("my"))
+        if self.cmd("add-member"):
+            self.add_member()
+        if self.cmd("remove-member"):
+            self.remove_member()
 
     def ls(self, my):
         try:
@@ -113,3 +126,93 @@ Options:{BASE_COMMAND_OPTIONS}
             ),
         )
         self.print("Team members", out)
+
+    def add_member(self):
+        try:
+            config = add_member_wizard(self.client)
+        except NoTeamError:
+            self.print(
+                "We haven't yet registered you as member of any Origo team. "
+                "Please contact Datapatruljen to get it done."
+            )
+            return
+
+        try:
+            user = self.client.get_user_by_username(config["username"])
+            team_members = [
+                m["username"] for m in self.client.get_team_members(config["team_id"])
+            ]
+
+            if user["username"] in team_members:
+                self.print(
+                    "User {} is already a member of this team.".format(
+                        self._team_member_representation(user)
+                    )
+                )
+                return
+
+            self.confirm_to_continue(
+                "Add {} to the team?".format(self._team_member_representation(user))
+            )
+
+            self.client.update_team_members(
+                config["team_id"], team_members + [user["username"]]
+            )
+            self.print("Done!")
+        except HTTPError as e:
+            message = e.response.json()["message"]
+            self.print(f"Something went wrong: {message}")
+            return
+
+    def remove_member(self):
+        try:
+            config = remove_member_wizard(self.client)
+        except NoTeamError:
+            self.print(
+                "We haven't yet registered you as member of any Origo team. "
+                "Please contact Datapatruljen to get it done."
+            )
+            return
+
+        try:
+            team_members = self.client.get_team_members(config["team_id"])
+            members_to_remove, members_to_keep = [], []
+            for member in team_members:
+                if member["username"] in config["usernames"]:
+                    members_to_remove.append(member)
+                else:
+                    members_to_keep.append(member)
+
+            if len(members_to_keep) == 0:
+                self.confirm_to_continue(
+                    "You are about to delete all members from the team, including "
+                    "yourself. It will not be possible to edit this team any further "
+                    "without being re-added by a systems administrator."
+                )
+            else:
+                self.confirm_to_continue(
+                    "Remove the following member{} from the team?\n - {}".format(
+                        "s" if len(members_to_remove) > 1 else "",
+                        "\n - ".join(
+                            [
+                                self._team_member_representation(m)
+                                for m in members_to_remove
+                            ]
+                        ),
+                    )
+                )
+
+            self.client.update_team_members(
+                config["team_id"], [m["username"] for m in members_to_keep]
+            )
+            self.print("Done!")
+        except HTTPError as e:
+            message = e.response.json()["message"]
+            self.print(f"Something went wrong: {message}")
+            return
+
+    @staticmethod
+    def _team_member_representation(user):
+        if user["name"]:
+            return f"{user['name']} ({user['username']})"
+        return user["username"]